chore: 주문 생성/조회 기능 및 관련 클래스 추가

+++ src/main/java/com/kt/service/product/ProductService.java
@@ -16,43 +16,43 @@ import lombok.RequiredArgsConstructor;
 public class ProductService {
 	private final ProductRepository repository;
 
-	public void create(String name, Long price, Long quantity){
+	public void create(String name, Long price, Long quantity) {
 		var product = Product.create(name, price, quantity);
 		repository.save(product);
 	}
 
-	public void update(Long id, String name, Long price, Long quantity){
+	public void update(Long id, String name, Long price, Long quantity) {
 		var product = repository.findByIdOrThrow(id);
 		product.update(name, price, quantity);
 	}
 
-	public void soldOut(Long id){
+	public void soldOut(Long id) {
 		var product = repository.findByIdOrThrow(id);
 		product.soldOut();
 	}
 
-	public void activate(Long id){
+	public void activate(Long id) {
 		var product = repository.findByIdOrThrow(id);
 		product.active();
 
 	}
 
-	public void inActive(Long id){
+	public void inActive(Long id) {
 		var product = repository.findByIdOrThrow(id);
 		product.inActive();
 	}
 
-	public void delete(Long id){
+	public void delete(Long id) {
 		var product = repository.findByIdOrThrow(id);
 		product.delete();
 	}
 
-	public void decreaseStock(Long id, int quantity){
+	public void decreaseStock(Long id, Long quantity) {
 		var product = repository.findByIdOrThrow(id);
 		product.decreaseStock(quantity);
 	}
 
-	public void increaseStock(Long id, int quantity){
+	public void increaseStock(Long id, Long quantity) {
 		var product = repository.findByIdOrThrow(id);
 		product.increaseStock(quantity);
 	}
